requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
backoff
bcrypt
pymongo 
motor
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
coloredlogs
cryptography
dataclasses-json
datasets
dill
distro
dnspython
durationpy
einops
email_validator
fastapi
filelock
filetype
flatbuffers
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
h2
hf-xet
hpack
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperframe
idna
importlib_metadata
importlib_resources
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-deepseek
langchain-google-genai
langchain-huggingface
langchain-openai
langchain-text-splitters
langchain-xai
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
motor
mpmath
multidict
multiprocess
mypy_extensions
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
ormsgpack
overrides
packaging
pandas
pillow
posthog
propcache
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
PyPika
pyproject_hooks
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
shellingham
six
sniffio
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
triton
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
xxhash
yarl
zipp
zstandard
gspread
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib